requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
argcomplete==2.0.0
asgiref==3.6.0
async-timeout==4.0.2
attrs==21.2.0
Automat==20.2.0
Babel==2.8.0
bcrypt==3.2.0
blinker==1.4
boto==2.49.0
cachetools==5.3.0
certifi==2020.6.20
chardet==4.0.0
charset-normalizer==3.0.1
click==8.0.3
cloud-init==23.1.2
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
crcmod==1.7
cryptography==3.4.8
dbus-python==1.2.18
distlib==0.3.6
distro==1.7.0
distro-info===1.1build1
Django==4.1.4
django-appconf==1.0.5
django-compressor==2.4
django-cors-headers==3.13.0
django3-livesync==1
djangorestframework==3.14.0
djangorestframework-jsonapi==6.0.0
fasteners==0.18
filelock==3.8.2
frozenlist==1.3.3
gcs-oauth2-boto-plugin==3.0
google-apitools==0.5.32
google-auth==2.16.2
google-reauth==0.1.1
gsutil==5.21
gunicorn==20.1.0
httplib2==0.20.4
humanize==4.4.0
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.6.4
incremental==21.3.0
inflection==0.5.1
jeepney==0.7.1
Jinja2==3.0.3
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
MarkupSafe==2.0.1
monotonic==1.6
more-itertools==8.10.0
multidict==6.0.4
netifaces==0.11.0
oauth2client==4.1.3
oauthlib==3.2.0
pexpect==4.8.0
Pillow==9.3.0
platformdirs==2.6.0
psycopg2==2.9.5
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.18.1
pyserial==3.5
python-apt==2.3.0+ubuntu2.1
python-debian===0.1.43ubuntu1
python-magic==0.4.24
pytz==2022.7
pyu2f==0.1.5
PyYAML==5.4.1
rcssmin==1.0.6
requests==2.25.1
retry-decorator==1.1.1
rjsmin==1.1.0
rsa==4.7.2
SecretStorage==3.3.1
service-identity==18.1.0
six==1.16.0
sos==4.4
sqlparse==0.4.3
ssh-import-id==5.11
supervisor==4.2.1
systemd-python==234
tornado==6.2
Twisted==22.1.0
ubuntu-advantage-tools==27.12
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
uuid==1.30
uWSGI==2.0.21
virtualenv==20.17.1
wadllib==1.3.6
watchdog==3.0.0
websocket-client==1.5.1
yarl==1.8.2
zipp==1.0.0
zope.interface==5.4.0